Seperate general type files from transaction package
