+++ core/signedTransaction.go
@@ -1,4 +1,4 @@
-package transaction
+package core
 
 import (
 	"encoding/hex"
